Factorise out Bluetooth module
