+++ plugin.xml
@@ -30,7 +30,7 @@
         <header-file src="src/ios/Diagnostic.h" />
         <source-file src="src/ios/Diagnostic.m" />
 
-        <framework src="CoreBluetooth.framework" />
+
         <framework src="CoreMotion.framework" />
         <framework src="AVFoundation.framework" />
         <framework src="Photos.framework" />
@@ -45,10 +45,6 @@
 
         <config-file target="*-Info.plist" parent="NSPhotoLibraryUsageDescription">
             <string>This app requires photo library access to function properly.</string>
-        </config-file>
-
-        <config-file target="*-Info.plist" parent="NSBluetoothPeripheralUsageDescription">
-            <string>This app requires bluetooth access to function properly.</string>
         </config-file>
 
         <config-file target="*-Info.plist" parent="NSMicrophoneUsageDescription">
@@ -102,6 +98,28 @@
         </config-file>
         <!--END_MODULE LOCATION-->
 
+        <!--BEGIN_MODULE BLUETOOTH-->
+        <config-file target="config.xml" parent="/*">
+            <feature name="Diagnostic_Bluetooth">
+                <param name="ios-package" value="Diagnostic_Bluetooth" />
+                <param name="onload" value="true" />
+            </feature>
+        </config-file>
+
+        <js-module src="www/ios/diagnostic.bluetooth.js" name="Diagnostic_Bluetooth">
+            <merges target="cordova.plugins.diagnostic.bluetooth" />
+        </js-module>
+
+        <header-file src="src/ios/Diagnostic_Bluetooth.h" />
+        <source-file src="src/ios/Diagnostic_Bluetooth.m" />
+
+        <framework src="CoreBluetooth.framework" />
+
+        <config-file target="*-Info.plist" parent="NSBluetoothPeripheralUsageDescription">
+            <string>This app requires bluetooth access to function properly.</string>
+        </config-file>
+        <!--END_MODULE BLUETOOTH-->
+
     </platform>
 
     <platform name="android">
@@ -153,6 +171,20 @@
         </config-file>
         <!--END_MODULE LOCATION-->
 
+        <!--BEGIN_MODULE BLUETOOTH-->
+        <config-file target="config.xml" parent="/*">
+            <feature name="Diagnostic_Bluetooth" >
+                <param name="android-package" value="cordova.plugins.Diagnostic_Bluetooth"/>
+                <param name="onload" value="true" />
+            </feature>
+        </config-file>
+
+        <js-module name="Diagnostic_Bluetooth" src="www/android/diagnostic.bluetooth.js">
+            <merges target="cordova.plugins.diagnostic.bluetooth" />
+        </js-module>
+
+        <source-file src="src/android/Diagnostic_Bluetooth.java" target-dir="src/cordova/plugins" />
+        <!--END_MODULE BLUETOOTH-->
     </platform>
 
     <platform name="windows">
@@ -187,5 +219,21 @@
             <merges target="" />
         </js-module>
         <!--END_MODULE LOCATION-->
+
+        <!--BEGIN_MODULE BLUETOOTH-->
+        <config-file target="config.xml" parent="/*">
+            <feature name="Diagnostic_Bluetooth" >
+                <param name="windows-package" value="cordova.plugins.Diagnostic_Bluetooth"/>
+            </feature>
+        </config-file>
+
+        <js-module src="www/windows/diagnostic.bluetooth.js" name="Diagnostic_Bluetooth">
+            <merges target="cordova.plugins.diagnostic.bluetooth" />
+        </js-module>
+
+        <js-module src="src/windows/diagnosticProxy.bluetooth.js" name="diagnosticProxy.bluetooth">
+            <merges target="" />
+        </js-module>
+        <!--END_MODULE BLUETOOTH-->
     </platform>
 </plugin>
