+++ src/ios/Diagnostic.m
@@ -192,11 +192,6 @@ ABAddressBookRef _addressBook;
     
     osVersion = [[[UIDevice currentDevice] systemVersion] floatValue];
     
-    self.bluetoothManager = [[CBCentralManager alloc]
-                             initWithDelegate:self
-                             queue:dispatch_get_main_queue()
-                             options:@{CBCentralManagerOptionShowPowerAlertKey: @(NO)}];
-    [self centralManagerDidUpdateState:self.bluetoothManager]; // Show initial state
     self.contactStore = [[CNContactStore alloc] init];
     self.motionManager = [[CMMotionActivityManager alloc] init];
     self.motionActivityQueue = [[NSOperationQueue alloc] init];
@@ -371,66 +366,6 @@ ABAddressBookRef _addressBook;
     return [cset countForObject:@"awdl0"] > 1 ? YES : NO;
 }
 
-
-#pragma mark - Bluetooth
-
-- (void) isBluetoothAvailable: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            NSString* state = [self getBluetoothState];
-            bool bluetoothEnabled;
-            if([state  isEqual: @"powered_on"]){
-                bluetoothEnabled = true;
-            }else{
-                bluetoothEnabled = false;
-            }
-            [self sendPluginResultBool:bluetoothEnabled :command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
-
-- (void) getBluetoothState: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            NSString* state = [self getBluetoothState];
-            [self logDebug:[NSString stringWithFormat:@"Bluetooth state is: %@", state]];
-            [self sendPluginResultString:state:command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-    
-}
-
-- (void) requestBluetoothAuthorization: (CDVInvokedUrlCommand*)command
-{
-    [self.commandDelegate runInBackground:^{
-        @try {
-            NSString* state = [self getBluetoothState];
-            
-            if([state isEqual: @"unauthorized"]){
-                /*
-                 When the application requests to start scanning for bluetooth devices that is when the user is presented with a consent dialog.
-                 */
-                [self logDebug:@"Requesting bluetooth authorization"];
-                [self.bluetoothManager scanForPeripheralsWithServices:nil options:nil];
-                [self.bluetoothManager stopScan];
-            }else{
-                [self logDebug:@"Bluetooth authorization is already granted"];
-            }
-            [self sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK] :command];
-        }
-        @catch (NSException *exception) {
-            [self handlePluginException:exception :command];
-        }
-    }];
-}
 
 #pragma mark -  Settings
 - (void) switchToSettings: (CDVInvokedUrlCommand*)command
@@ -1305,79 +1240,6 @@ ABAddressBookRef _addressBook;
 }
 #endif
 
-
-- (NSString*)getBluetoothState{
-    NSString* state;
-    NSString* description;
-    
-    switch(self.bluetoothManager.state)
-    {
-            
-#if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
-        case CBManagerStateResetting:
-#else
-        case CBCentralManagerStateResetting:
-#endif
-            state = @"resetting";
-            description =@"The connection with the system service was momentarily lost, update imminent.";
-            break;
-            
-#if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
-        case CBManagerStateUnsupported:
-#else
-        case CBCentralManagerStateUnsupported:
-#endif
-            state = @"unsupported";
-            description = @"The platform doesn't support Bluetooth Low Energy.";
-            break;
-            
-#if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
-        case CBManagerStateUnauthorized:
-#else
-        case CBCentralManagerStateUnauthorized:
-#endif
-            state = @"unauthorized";
-            description = @"The app is not authorized to use Bluetooth Low Energy.";
-            break;
-            
-#if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
-        case CBManagerStatePoweredOff:
-#else
-        case CBCentralManagerStatePoweredOff:
-#endif
-            state = @"powered_off";
-            description = @"Bluetooth is currently powered off.";
-            break;
-            
-#if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
-        case CBManagerStatePoweredOn:
-#else
-        case CBCentralManagerStatePoweredOn:
-#endif
-            state = @"powered_on";
-            description = @"Bluetooth is currently powered on and available to use.";
-            break;
-        default:
-            state = UNKNOWN;
-            description = @"State unknown, update imminent.";
-            break;
-    }
-    [self logDebug:[NSString stringWithFormat:@"Bluetooth state changed: %@",description]];
-    
-    
-    return state;
-}
-
-/********************************/
-#pragma mark - CBCentralManagerDelegate
-/********************************/
-
-- (void) centralManagerDidUpdateState:(CBCentralManager *)central {
-    
-    NSString* state = [self getBluetoothState];
-    NSString* jsString = [NSString stringWithFormat:@"cordova.plugins.diagnostic._onBluetoothStateChange(\"%@\");", state];
-    [self executeGlobalJavascript:jsString];
-}
 
 @end
